+++ rust/analytics/tests/sql_view_test.rs
@@ -17,7 +17,9 @@ use micromegas_analytics::lakehouse::write_partition::{RetireMatch, retire_parti
 use micromegas_analytics::lakehouse::{sql_batch_view::SqlBatchView, view_factory::ViewFactory};
 use micromegas_analytics::response_writer::{Logger, ResponseWriter};
 use micromegas_analytics::time::TimeRange;
-use micromegas_ingestion::data_lake_connection::{DataLakeConnection, connect_to_data_lake};
+use micromegas_ingestion::data_lake_connection::{
+    DataLakeConnection, WritablePolicy, connect_to_data_lake,
+};
 use micromegas_telemetry_sink::TelemetryGuardBuilder;
 use micromegas_tracing::prelude::*;
 use std::sync::Arc;
@@ -324,7 +326,14 @@ async fn sql_view_test() -> Result<()> {
     let object_store_uri = std::env::var("MICROMEGAS_OBJECT_STORE_URI")
         .with_context(|| "reading MICROMEGAS_OBJECT_STORE_URI")?;
     let runtime = Arc::new(make_runtime_env()?);
-    let lake = Arc::new(connect_to_data_lake(&connection_string, &object_store_uri).await?);
+    let lake = Arc::new(
+        connect_to_data_lake(
+            WritablePolicy::Require,
+            &connection_string,
+            &object_store_uri,
+        )
+        .await?,
+    );
     let default_audience_lakehouse =
         Arc::new(LakehouseContext::new(lake.clone(), runtime.clone())?);
     let log_summary_view = Arc::new(